fix: #236 — Organization.balance Int→Long, миграция BIGINT
balance: Int (max ~21.4 млн руб) переполнялся при зачислении выручки
крупным организациям, вызывая IllegalArgumentException на легитимных
транзакциях. Заменён на Long (~9.2 * 10^18) во всех слоях.

- Organization.balance: Long = 0L
- JdbcOrganizationRepository: rs.getLong("balance")
- V25__organization_balance_bigint.sql: ALTER COLUMN balance TYPE BIGINT
- OrganizationCreditOverflowTests: обновлены типы для Long

Closes #236

+++ src/main/kotlin/com/karrad/bilets/domain/entity/Organization.kt
@@ -5,13 +5,13 @@ import java.util.UUID
 data class Organization(
     val code: String,
     val name: String,
-    val balance: Int = 0,
+    val balance: Long = 0L,
     val id: UUID = UUID.randomUUID()
 ) {
     init {
         require(code.isNotBlank()) { "Organization code must not be blank" }
         require(name.isNotBlank()) { "Organization name must not be blank" }
-        require(balance >= 0) { "Organization balance must not be negative" }
+        require(balance >= 0L) { "Organization balance must not be negative" }
     }
 
     fun credit(amount: Int): Organization {
